requirement.txt
python-dotenv
bootstrap-flask
flask-sqlalchemy
flask-wtf 
flask-login
pyyaml